Add Equinox Sky generator
For LEED EQc7 Option 2, a sky configuration at equinox is needed.

The illuminance intensity for sun and sky are taken from the clearest sky conditions within 15 days of September 21 and March 21. The clearness definition is taken from Perez sky theory.

A new gendaylit command is added, in order to generate a sky direct and diffuse components from the weather file.