LEED Results
For the justificacion of Credit EQc7 for LEED about Daylight is interesting to know which percentage of the area is below 300 lux and which is above 3000 lux.

For this reason, in the results is added the computation of the percentage of area that is out of the limits of the Objective.